Revert to unsigned int

+++ src/Distance/Distance.hpp
@@ -128,82 +128,93 @@ namespace devuino
 	/* User defined litterals */
 
 	constexpr Distance operator""_gigametre(const long double lenght) { return Distance {lenght * 1000000000.0}; }
-	constexpr Distance operator""_gigametre(const long long int lenght) { return Distance {lenght * 1000000000.0}; }
+	constexpr Distance operator""_gigametre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000000000.0}; }
+
 	constexpr Distance operator""_Gm(const long double lenght) { return Distance {lenght * 1000000000.0}; }
-	constexpr Distance operator""_Gm(const long long int lenght) { return Distance {lenght * 1000000000.0}; }
+	constexpr Distance operator""_Gm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000000000.0}; }
 
 	constexpr Distance operator""_megametre(const long double lenght) { return Distance {lenght * 1000000.0}; }
-	constexpr Distance operator""_megametre(const long long int lenght) { return Distance {lenght * 1000000.0}; }
+	constexpr Distance operator""_megametre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000000.0}; }
+
 	constexpr Distance operator""_Mm(const long double lenght) { return Distance {lenght * 1000000.0}; }
-	constexpr Distance operator""_Mm(const long long int lenght) { return Distance {lenght * 1000000.0}; }
+	constexpr Distance operator""_Mm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000000.0}; }
 
 	constexpr Distance operator""_kilometre(const long double lenght) { return Distance {lenght * 1000.0}; }
-	constexpr Distance operator""_kilometre(const long long int lenght) { return Distance {lenght * 1000.0}; }
+	constexpr Distance operator""_kilometre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000.0}; }
+
 	constexpr Distance operator""_km(const long double lenght) { return Distance {lenght * 1000.0}; }
-	constexpr Distance operator""_km(const long long int lenght) { return Distance {lenght * 1000.0}; }
+	constexpr Distance operator""_km(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1000.0}; }
 
 	constexpr Distance operator""_hektometre(const long double lenght) { return Distance {lenght * 100.0}; }
-	constexpr Distance operator""_hektometre(const long long int lenght) { return Distance {lenght * 100.0}; }
+	constexpr Distance operator""_hektometre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 100.0}; }
+
 	constexpr Distance operator""_hm(const long double lenght) { return Distance {lenght * 100.0}; }
-	constexpr Distance operator""_hm(const long long int lenght) { return Distance {lenght * 100.0}; }
+	constexpr Distance operator""_hm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 100.0}; }
 
 	constexpr Distance operator""_dekametre(const long double lenght) { return Distance {lenght * 10.0}; }
-	constexpr Distance operator""_dekametre(const long long int lenght) { return Distance {lenght * 10.0}; }
+	constexpr Distance operator""_dekametre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 10.0}; }
+
 	constexpr Distance operator""_dam(const long double lenght) { return Distance {lenght * 10.0}; }
-	constexpr Distance operator""_dam(const long long int lenght) { return Distance {lenght * 10.0}; }
+	constexpr Distance operator""_dam(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 10.0}; }
 
 	constexpr Distance operator""_metre(const long double lenght) { return Distance {lenght}; }
-	constexpr Distance operator""_metre(const long long int lenght) { return Distance {lenght}; }
+	constexpr Distance operator""_metre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght)}; }
+
 	constexpr Distance operator""_m(const long double lenght) { return Distance {lenght}; }
-	constexpr Distance operator""_m(const long long int lenght) { return Distance {lenght}; }
+	constexpr Distance operator""_m(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght)}; }
 
 	constexpr Distance operator""_decimetre(const long double lenght) { return Distance {lenght / 10.0}; }
-	constexpr Distance operator""_decimetre(const long long int lenght) { return Distance {lenght / 10.0}; }
+	constexpr Distance operator""_decimetre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 10.0}; }
+
 	constexpr Distance operator""_dm(const long double lenght) { return Distance {lenght / 10.0}; }
-	constexpr Distance operator""_dm(const long long int lenght) { return Distance {lenght / 10.0}; }
+	constexpr Distance operator""_dm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 10.0}; }
 
 	constexpr Distance operator""_centimetre(const long double lenght) { return Distance {lenght / 100.0}; }
-	constexpr Distance operator""_centimetre(const long long int lenght) { return Distance {lenght / 100.0}; }
+	constexpr Distance operator""_centimetre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 100.0}; }
+
 	constexpr Distance operator""_cm(const long double lenght) { return Distance {lenght / 100.0}; }
-	constexpr Distance operator""_cm(const long long int lenght) { return Distance {lenght / 100.0}; }
+	constexpr Distance operator""_cm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 100.0}; }
 
 	constexpr Distance operator""_millimetre(const long double lenght) { return Distance {lenght / 1000.0}; }
-	constexpr Distance operator""_millimetre(const long long int lenght) { return Distance {lenght / 1000.0}; }
+	constexpr Distance operator""_millimetre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000.0}; }
+
 	constexpr Distance operator""_mm(const long double lenght) { return Distance {lenght / 1000.0}; }
-	constexpr Distance operator""_mm(const long long int lenght) { return Distance {lenght / 1000.0}; }
+	constexpr Distance operator""_mm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000.0}; }
 
 	constexpr Distance operator""_micrometre(const long double lenght) { return Distance {lenght / 1000000.0}; }
-	constexpr Distance operator""_micrometre(const long long int lenght) { return Distance {lenght / 1000000.0}; }
+	constexpr Distance operator""_micrometre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000000.0}; }
+
 	constexpr Distance operator""_um(const long double lenght) { return Distance {lenght / 1000000.0}; }
-	constexpr Distance operator""_um(const long long int lenght) { return Distance {lenght / 1000000.0}; }
+	constexpr Distance operator""_um(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000000.0}; }
 
 	constexpr Distance operator""_nanometre(const long double lenght) { return Distance {lenght / 1000000000.0}; }
-	constexpr Distance operator""_nanometre(const long long int lenght) { return Distance {lenght / 1000000000.0}; }
+	constexpr Distance operator""_nanometre(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000000000.0}; }
+
 	constexpr Distance operator""_nm(const long double lenght) { return Distance {lenght / 1000000000.0}; }
-	constexpr Distance operator""_nm(const long long int lenght) { return Distance {lenght / 1000000000.0}; }
+	constexpr Distance operator""_nm(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) / 1000000000.0}; }
 
 	constexpr Distance operator""_thou(const long double lenght) { return Distance {lenght * 0.0000254}; }
-	constexpr Distance operator""_thou(const long long int lenght) { return Distance {lenght * 0.0000254}; }
+	constexpr Distance operator""_thou(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 0.0000254}; }
 
 	constexpr Distance operator""_inch(const long double lenght) { return Distance {lenght * 0.0254}; }
-	constexpr Distance operator""_inch(const long long int lenght) { return Distance {lenght * 0.0254}; }
+	constexpr Distance operator""_inch(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 0.0254}; }
 
 	constexpr Distance operator""_foot(const long double lenght) { return Distance {lenght * 0.3048}; }
-	constexpr Distance operator""_foot(const long long int lenght) { return Distance {lenght * 0.3048}; }
+	constexpr Distance operator""_foot(const unsigned long long int lenght) { return Distance {lenght * 0.3048}; }
 
 	constexpr Distance operator""_yard(const long double lenght) { return Distance {lenght * 0.9144}; }
-	constexpr Distance operator""_yard(const long long int lenght) { return Distance {lenght * 0.9144}; }
+	constexpr Distance operator""_yard(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 0.9144}; }
 
 	constexpr Distance operator""_chain(const long double lenght) { return Distance {lenght * 20.1168}; }
-	constexpr Distance operator""_chain(const long long int lenght) { return Distance {lenght * 20.1168}; }
+	constexpr Distance operator""_chain(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 20.1168}; }
 
 	constexpr Distance operator""_furlong(const long double lenght) { return Distance {lenght * 201.168}; }
-	constexpr Distance operator""_furlong(const long long int lenght) { return Distance {lenght * 201.168}; }
+	constexpr Distance operator""_furlong(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 201.168}; }
 
 	constexpr Distance operator""_mile(const long double lenght) { return Distance {lenght * 1609.344}; }
-	constexpr Distance operator""_mile(const long long int lenght) { return Distance {lenght * 1609.344}; }
+	constexpr Distance operator""_mile(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1609.344}; }
 
 	constexpr Distance operator""_nautical(const long double lenght) { return Distance {lenght * 1852.0}; }
-	constexpr Distance operator""_nautical(const long long int lenght) { return Distance {lenght * 1852.0}; }
+	constexpr Distance operator""_nautical(const unsigned long long int lenght) { return Distance {static_cast<double>(lenght) * 1852.0}; }
 
 }
